Added values from the resurrection_waiting_times table Added DeathDurabilityLossRatio, now the character's equipment durability will be reduced when he dies.
Added buff Recent Defeat when a character dies
Now the correct resurrection time from resurrection_waiting_times will be specified on death
Added experience loss on death
Added experience recovery after death

Now when a mob uses NpcReturn, it will reset Combat
Revamped the targeting system for mobs (Now a mob will take the most damage that was dealt to it). When a character dies, the mob will target the next character that comes after the killed character in terms of damage.
Also, if another character did more damage to the mob than you did, the mob will switch the targeting to it.

Fixed pet death (Now when a mount dies, it has Injury(1578) and Tripped(1579) buffs, which does not allow it to regenerate health, it must be resurrected through a special NPC).
Added the ability to raise a mount if it dies (just removes the Tripped(1579) buff).
Added ability to resurrect a pet with - Pet Dual-Recovery Potion(18649)
Fixed display of experience in the pet info window (now it is displayed in real time)

Fixed Slave death
Now when a Slave is destroyed, it will be correctly recalled
Added ability to repair Slave with Shatigon's Sandglass.